backend/python/requirements.txt
# Tooth Kingdom Adventure - Python Backend Dependencies
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
passlib>=1.7.4
bcrypt>=4.0.0
PyJWT>=2.8.0
google-generativeai>=0.8.3
elevenlabs>=1.0.0
python-dotenv>=1.0.1
requests>=2.32.0
colorama>=0.4.6
